# backend/openapi.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LiveResponse": {
        "properties": {
          "service": {
            "const": "resale-monitor",
            "default": "resale-monitor",
            "title": "Service",
            "type": "string"
          },
          "status": {
            "const": "ok",
            "default": "ok",
            "title": "Status",
            "type": "string"
          }
        },
        "title": "LiveResponse",
        "type": "object"
      },
      "ReadinessChecks": {
        "properties": {
          "database": {
            "const": "ok",
            "default": "ok",
            "title": "Database",
            "type": "string"
          }
        },
        "title": "ReadinessChecks",
        "type": "object"
      },
      "ReadinessResponse": {
        "properties": {
          "checks": {
            "$ref": "#/components/schemas/ReadinessChecks"
          },
          "service": {
            "const": "resale-monitor",
            "default": "resale-monitor",
            "title": "Service",
            "type": "string"
          },
          "status": {
            "const": "ok",
            "default": "ok",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "checks"
        ],
        "title": "ReadinessResponse",
        "type": "object"
      },
      "SearchScopeRead": {
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "geography": {
            "additionalProperties": true,
            "title": "Geography",
            "type": "object"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "purpose": {
            "title": "Purpose",
            "type": "string"
          },
          "query": {
            "additionalProperties": true,
            "title": "Query",
            "type": "object"
          }
        },
        "required": [
          "id",
          "provider",
          "purpose",
          "query",
          "geography",
          "enabled"
        ],
        "title": "SearchScopeRead",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WatchlistCreate": {
        "properties": {
          "center_place": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Center Place",
            "type": "string"
          },
          "maximum_price_minor": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maximum Price Minor"
          },
          "minimum_price_minor": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minimum Price Minor"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "query": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Query",
            "type": "string"
          },
          "radius_miles": {
            "exclusiveMinimum": 0.0,
            "maximum": 500.0,
            "title": "Radius Miles",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "query",
          "center_place",
          "radius_miles"
        ],
        "title": "WatchlistCreate",
        "type": "object"
      },
      "WatchlistRead": {
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "center_place": {
            "title": "Center Place",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "maximum_price_minor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maximum Price Minor"
          },
          "minimum_price_minor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minimum Price Minor"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "radius_miles": {
            "title": "Radius Miles",
            "type": "integer"
          },
          "scopes": {
            "items": {
              "$ref": "#/components/schemas/SearchScopeRead"
            },
            "title": "Scopes",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "category",
          "status",
          "center_place",
          "radius_miles",
          "minimum_price_minor",
          "maximum_price_minor",
          "currency",
          "created_at",
          "scopes"
        ],
        "title": "WatchlistRead",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Resale Monitor API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/health": {
      "get": {
        "operationId": "readiness_api_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadinessResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Readiness",
        "tags": [
          "health"
        ]
      }
    },
    "/api/health/live": {
      "get": {
        "operationId": "liveness_api_health_live_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LiveResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Liveness",
        "tags": [
          "health"
        ]
      }
    },
    "/api/watchlists": {
      "get": {
        "operationId": "list_all_api_watchlists_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WatchlistRead"
                  },
                  "title": "Response List All Api Watchlists Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List All",
        "tags": [
          "watchlists"
        ]
      },
      "post": {
        "operationId": "create_api_watchlists_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchlistCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create",
        "tags": [
          "watchlists"
        ]
      }
    },
    "/api/watchlists/{watchlist_id}": {
      "get": {
        "operationId": "detail_api_watchlists__watchlist_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "watchlist_id",
            "required": true,
            "schema": {
              "title": "Watchlist Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Detail Api Watchlists  Watchlist Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Detail",
        "tags": [
          "watchlists"
        ]
      }
    },
    "/api/watchlists/{watchlist_id}/runs": {
      "post": {
        "operationId": "run_api_watchlists__watchlist_id__runs_post",
        "parameters": [
          {
            "in": "path",
            "name": "watchlist_id",
            "required": true,
            "schema": {
              "title": "Watchlist Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "integer"
                  },
                  "title": "Response Run Api Watchlists  Watchlist Id  Runs Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run",
        "tags": [
          "watchlists"
        ]
      }
    }
  }
}